Change the return type of `FromStrFormat::regex` from `&str` to `String`.
If the return type is `str`, `FromStrFormat` must own a string, but `FromStrFormat` is instantiated each time `FromStr::from_str` is called, which reduces performance.

On the other hand, if the return type is `String`, there is an overhead to create `String` even if the regex is constant, but since the regex is cached for each field, there is almost no performance impact.