Split `FromStrFormat` into `FromStrFormat` and `FromStrFormatBase` to prevent incorrect regular expressions from being cached.
If the field annotated with `#[display(with = ...)]` included type parameters, the result of `FromStrFormat::regex` could vary depending on the type parameters. However, since regular expressions were cached under the assumption that they were the same regardless of type parameters, there was a possibility of incorrect regular expressions being cached.

By splitting `FromStrFormat` into `FromStrFormat` and `FromStrFormatBase`, when retrieving the regular expression, `FromStrFormatBase` is obtained without inferring the field's type, forcing the retrieval of a regular expression that does not depend on the field's type.